+++ src/OpenTelemetry.Exporter.Geneva/CHANGELOG.md
@@ -2,14 +2,21 @@
 
 ## Unreleased
 
-* Update OTel SDK version to `1.4.0-beta.1`.
 * Add support for exporting `UpDownCounter` and `ObservableUpDownCounter`.
   [#685](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/685)
+## 1.4.0-beta.2
+
+Released 2022-Oct-17
 
 * The option `TableNameMappings` of `GenevaExporterOptions` will not support
   string values that are null, empty, or consist only of white-space characters.
   It will also not support string values that contain non-ASCII characters.
   [646](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/646)
+
+* Update OTel SDK version to `1.4.0-beta.2`. Add support for exporting Histogram
+  Min and Max. If the histogram does not contain min and max, the exporter
+  exports both the values as zero.
+  [#704](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/704)
 
 ## 1.4.0-beta.1